[Sema] Hide the implementation of AccessScopeChecker
(and UsableFromInlineChecker)

Preparation for the next commit's refactoring
